requirements.txt
# Core ASR
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0

# NLP and ML
transformers==4.36.0
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.7.2
langdetect==1.0.9

# Data handling
pandas==2.1.3
numpy==1.24.3
tqdm==4.66.1

# Evaluation
jiwer==3.0.3
rouge-score==0.1.2
bert-score==0.3.13  # Fixed version
evaluate==0.4.1

# Development
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0

# Web interface
streamlit==1.28.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Optional TTS for testing
gtts==2.5.0